src/requirements.txt
kafka-python == 2.0.2
boto3 == 1.35.18
s3fs == 2024.9.0
psycopg2 == 3.2.1
python-dotenv == 1.0.1
pydantic_settings == 2.3.4